pyproject.toml
[tool.poetry]
name = "esp32-camera-station"
version = "0.1.0"
description = ""
authors = ["kohei <koheitech001@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "esp32_camera_station"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
